fix(mempool price): created global variable in context
