+++ apps/web/src/providers/context.tsx
@@ -16,6 +16,11 @@ interface Session {
   user: SessionData;
 }
 
+interface MempoolPrice {
+  USD: number;
+  EUR: number;
+}
+
 interface AppContext {
   // User
   user: UserDetails | null;
@@ -36,6 +41,10 @@ interface AppContext {
   // Blog
   blogs: JoinedBlogLight[] | null;
   setBlogs: (blogs: JoinedBlogLight[] | null) => void;
+
+  // Conversion rate
+  conversionRate: number | null;
+  setConversionRate: (rate: number | null) => void;
 }
 
 export const AppContext = createContext<AppContext>({
@@ -58,6 +67,10 @@ export const AppContext = createContext<AppContext>({
   // Blog
   blogs: null,
   setBlogs: () => {},
+
+  // Conversion Rate
+  conversionRate: null,
+  setConversionRate: () => {},
 });
 
 export const AppContextProvider = ({ children }: PropsWithChildren) => {
@@ -70,6 +83,30 @@ export const AppContextProvider = ({ children }: PropsWithChildren) => {
   );
   const [courses, setCourses] = useState<JoinedCourse[] | null>(null);
   const [blogs, setBlogs] = useState<JoinedBlogLight[] | null>(null);
+  const [conversionRate, setConversionRate] = useState<number | null>(null);
+
+  const fetchConversionRate = async (): Promise<void> => {
+    try {
+      const response = await fetch('https://mempool.space/api/v1/prices');
+      const data: MempoolPrice = await response.json();
+
+      if (data?.USD) {
+        setConversionRate(data.USD);
+      } else {
+        throw new Error('Failed to retrieve conversion rate from Kraken API.');
+      }
+    } catch (error) {
+      console.error('Error fetching conversion rate:', error);
+    }
+  };
+
+  useEffect(() => {
+    fetchConversionRate();
+
+    const interval = setInterval(fetchConversionRate, 5 * 60 * 1000); // 5 minutes
+
+    return () => clearInterval(interval);
+  }, []);
 
   useEffect(() => {
     trpcClient.user.getDetails
@@ -128,6 +165,8 @@ export const AppContextProvider = ({ children }: PropsWithChildren) => {
     setCourses,
     blogs,
     setBlogs,
+    conversionRate,
+    setConversionRate,
   };
 
   return (
