+++ apps/web/src/routes/_content/_misc/-components/b-certificate-events.tsx
@@ -19,7 +19,7 @@ interface BCertificateEventsProps {
 }
 
 export const BCertificateEvents = ({ events }: BCertificateEventsProps) => {
-  const { session } = useContext(AppContext);
+  const { session, conversionRate } = useContext(AppContext);
   const isLoggedIn = !!session;
 
   const { data: eventPayments, refetch: refetchEventPayments } =
@@ -37,8 +37,6 @@ export const BCertificateEvents = ({ events }: BCertificateEventsProps) => {
       accessType: null,
     });
   const [isPaymentModalOpen, setIsPaymentModalOpen] = useState(false);
-
-  const [conversionRate, setConversionRate] = useState<number | null>(null);
 
   const payingEvent: JoinedEvent | undefined = events?.find(
     (e) => e.id === paymentModalData.eventId,
@@ -68,31 +66,6 @@ export const BCertificateEvents = ({ events }: BCertificateEventsProps) => {
     isOpen: isAuthModalOpen,
     close: closeAuthModal,
   } = useDisclosure();
-
-  interface MempoolPrice {
-    USD: number;
-    EUR: number;
-  }
-
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const response = await fetch('https://mempool.space/api/v1/prices');
-        const data = (await response.json()) as MempoolPrice;
-
-        if (data) {
-          const newConversionRate = data.USD;
-          setConversionRate(newConversionRate);
-        } else {
-          console.error('Failed to retrieve conversion rate from Kraken API.');
-        }
-      } catch (error) {
-        console.error('Failed to fetch conversion rate:', error);
-      }
-    }
-
-    fetchData();
-  }, []);
 
   // TODO refactor prop drilling
   return (
