+++ apps/web/src/routes/_content/resources/-components/builder-events.tsx
@@ -19,7 +19,7 @@ interface BuilderEventsProps {
 }
 
 export const BuilderEvents = ({ events }: BuilderEventsProps) => {
-  const { session } = useContext(AppContext);
+  const { session, conversionRate } = useContext(AppContext);
   const isLoggedIn = !!session;
   useEffect(() => {
     console.log('Events data:', events);
@@ -41,8 +41,6 @@ export const BuilderEvents = ({ events }: BuilderEventsProps) => {
       accessType: null,
     });
   const [isPaymentModalOpen, setIsPaymentModalOpen] = useState(false);
-
-  const [conversionRate, setConversionRate] = useState<number | null>(null);
 
   const payingEvent: JoinedEvent | undefined = events?.find(
     (e) => e.id === paymentModalData.eventId,
@@ -72,31 +70,6 @@ export const BuilderEvents = ({ events }: BuilderEventsProps) => {
     isOpen: isAuthModalOpen,
     close: closeAuthModal,
   } = useDisclosure();
-
-  interface MempoolPrice {
-    USD: number;
-    EUR: number;
-  }
-
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const response = await fetch('https://mempool.space/api/v1/prices');
-        const data = (await response.json()) as MempoolPrice;
-
-        if (data) {
-          const newConversionRate = data.USD;
-          setConversionRate(newConversionRate);
-        } else {
-          console.error('Failed to retrieve conversion rate from Kraken API.');
-        }
-      } catch (error) {
-        console.error('Failed to fetch conversion rate:', error);
-      }
-    }
-
-    fetchData();
-  }, []);
 
   // TODO refactor prop drilling
   return (
